+++ pkg/p4rtclient/conn_manager_test.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package southbound
+package p4rtclient
 
 import (
 	"context"
